requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.1.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
